Polish spacing & colors.
