+++ packages/apps/plugins/plugin-files/src/FilesPlugin.tsx
@@ -169,6 +169,7 @@ export const FilesPlugin = (): PluginDefinition<LocalFilesPluginProvides, Markdo
           const [groupNode] = parent.add({
             id: 'all-files',
             label: ['plugin name', { ns: FILES_PLUGIN }],
+            properties: { palette: 'teal' },
           });
 
           const fileIndices = getIndices(state.files.length);
