+++ packages/apps/plugins/plugin-space/src/SpacePlugin.tsx
@@ -157,6 +157,7 @@ export const SpacePlugin = (): PluginDefinition<SpacePluginProvides> => {
           const [groupNode] = parent.add({
             id: getSpaceId('all-spaces'),
             label: ['plugin name', { ns: SPACE_PLUGIN }],
+            properties: { palette: 'blue' },
           });
 
           const client = clientPlugin.provides.client;
