+++ packages/apps/plugins/plugin-treeview/src/components/NavTree/CollapsibleHeading.tsx
@@ -7,7 +7,7 @@ import React from 'react';
 
 import { Graph } from '@braneframe/plugin-graph';
 import { TreeItem, useSidebars, useTranslation } from '@dxos/aurora';
-import { getSize, mx, staticDisabled, valenceColorText } from '@dxos/aurora-theme';
+import { getSize, ghostButtonColors, mx, staticDisabled, valenceColorText } from '@dxos/aurora-theme';
 
 import { TREE_VIEW_PLUGIN } from '../../types';
 import {
@@ -17,7 +17,7 @@ import {
   topLevelHeadingColor,
   topLevelText,
   treeItemText,
-} from './style-fragments';
+} from './navtree-fragments';
 
 export const CollapsibleHeading = ({ open, node, level }: { open: boolean; level: number; node: Graph.Node }) => {
   const { navigationSidebarOpen } = useSidebars();
@@ -44,7 +44,7 @@ export const CollapsibleHeading = ({ open, node, level }: { open: boolean; level
     <TreeItem.OpenTrigger
       {...(disabled && { disabled, 'aria-disabled': true })}
       {...(!navigationSidebarOpen && { tabIndex: -1 })}
-      classNames={['grow flex items-center gap-1 pie-1', disabled && staticDisabled]}
+      classNames={['grow flex items-center gap-1 pie-1', ghostButtonColors, disabled && staticDisabled]}
     >
       <OpenTriggerIcon weight='fill' className={mx('shrink-0', getSize(2))} />
       {node.icon && <node.icon className={getSize(4)} />}
