+++ packages/apps/plugins/plugin-treeview/src/components/NavTree/NavTreeItem.tsx
@@ -10,14 +10,15 @@ import React, { FC, forwardRef, ForwardRefExoticComponent, RefAttributes, useEff
 import { SortableProps } from '@braneframe/plugin-dnd';
 import { Graph } from '@braneframe/plugin-graph';
 import { Button, DropdownMenu, Tooltip, TreeItem, useSidebars, useTranslation } from '@dxos/aurora';
-import { focusRing, getSize } from '@dxos/aurora-theme';
+import { focusRing, getSize, groupSurface, mx } from '@dxos/aurora-theme';
 
+import { useTreeView } from '../../TreeViewContext';
 import { SharedTreeItemProps, TREE_VIEW_PLUGIN } from '../../types';
 import { sortActions } from '../../util';
 import { CollapsibleHeading } from './CollapsibleHeading';
 import { NavTree } from './NavTree';
 import { NavigableHeading } from './NavigableHeading';
-import { levelPadding } from './style-fragments';
+import { levelPadding } from './navtree-fragments';
 
 type SortableBranchTreeViewItemProps = SharedTreeItemProps & Pick<SortableProps, 'rearranging'>;
 
@@ -54,6 +55,7 @@ export const NavTreeItem: ForwardRefExoticComponent<TreeViewItemProps & RefAttri
   const actions = sortActions(node.actions);
   const { t } = useTranslation(TREE_VIEW_PLUGIN);
   const { navigationSidebarOpen } = useSidebars();
+  const { active: treeViewActive } = useTreeView();
 
   const suppressNextTooltip = useRef<boolean>(false);
   const [optionsTooltipOpen, setOptionsTooltipOpen] = useState(false);
@@ -62,6 +64,7 @@ export const NavTreeItem: ForwardRefExoticComponent<TreeViewItemProps & RefAttri
   const [open, setOpen] = useState(level < 1);
 
   const disabled = !!node.properties?.disabled;
+  const active = treeViewActive === node.id;
 
   useEffect(() => {
     // todo(thure): Open if child within becomes active
@@ -72,13 +75,13 @@ export const NavTreeItem: ForwardRefExoticComponent<TreeViewItemProps & RefAttri
       collapsible={isBranch}
       open={!disabled && open}
       onOpenChange={(nextOpen) => setOpen(disabled ? false : nextOpen)}
-      classNames={['rounded block', focusRing, levelPadding(level), rearranging && 'invisible']}
+      classNames={['rounded block', focusRing, active && groupSurface, rearranging && 'invisible']}
       {...draggableAttributes}
       {...draggableListeners}
       style={style}
       ref={forwardedRef}
     >
-      <div role='none' className='flex items-start pie-1'>
+      <div role='none' className={mx(levelPadding(level), 'flex items-start')}>
         {isBranch ? (
           <CollapsibleHeading open={open} node={node} level={level} />
         ) : (
